* origin/master: (243 commits)
  Be consistent in referring to cori as cori-knl
  Add test for project=E3SM so that a non-cime build isn't confused
  Only test on USE_CUDA as the other variable is not set
  Some changes to build on Perlmutter with gnu v10
  Change enable checks =1 instead of true
  Remove temporary override of tensor HV setting for ne4
  SCREAM: fix valgrind error related to homme's readnl
  Fix param name in input.yaml
  Set tracers to 10 and enabled field checking
  Fix more YAKL includes
  Fix YAKL includes
  Fix scorpio cmake settings for scream
  Set default PIO output type to be 64bit
  Use size_t in another location
  Adding another change to use size_t
  Try changing a few types that represent sizes to size_t
  SCREAM: fixed comments/messages in spa
  Make default timestep 5min for SCREAMv1 CIME cases (#1489)
  SCREAM: final fix to spa_main test setup
  SCREAM: fix creation of SPADataHost in spa_main_test
  ...

+++ components/scream/src/physics/shoc/atmosphere_macrophysics.cpp
@@ -117,23 +117,23 @@ set_computed_group_impl (const FieldGroup& group)
 }
 
 // =========================================================================================
-int SHOCMacrophysics::requested_buffer_size_in_bytes() const
+size_t SHOCMacrophysics::requested_buffer_size_in_bytes() const
 {
   const int nlev_packs       = ekat::npack<Spack>(m_num_levs);
   const int nlevi_packs      = ekat::npack<Spack>(m_num_levs+1);
   const int num_tracer_packs = ekat::npack<Spack>(m_num_tracers);
 
   // Number of Reals needed by local views in the interface
-  const int interface_request = Buffer::num_1d_scalar*m_num_cols*sizeof(Real) +
-                                Buffer::num_2d_vector_mid*m_num_cols*nlev_packs*sizeof(Spack) +
-                                Buffer::num_2d_vector_int*m_num_cols*nlevi_packs*sizeof(Spack) +
-                                Buffer::num_2d_vector_tr*m_num_cols*num_tracer_packs*sizeof(Spack);
+  const size_t interface_request = Buffer::num_1d_scalar*m_num_cols*sizeof(Real) +
+                                   Buffer::num_2d_vector_mid*m_num_cols*nlev_packs*sizeof(Spack) +
+                                   Buffer::num_2d_vector_int*m_num_cols*nlevi_packs*sizeof(Spack) +
+                                   Buffer::num_2d_vector_tr*m_num_cols*num_tracer_packs*sizeof(Spack);
 
   // Number of Reals needed by the WorkspaceManager passed to shoc_main
   const auto policy       = ekat::ExeSpaceUtils<KT::ExeSpace>::get_default_team_policy(m_num_cols, nlev_packs);
   const int n_wind_slots  = ekat::npack<Spack>(2)*Spack::n;
   const int n_trac_slots  = ekat::npack<Spack>(m_num_tracers+3)*Spack::n;
-  const int wsm_request   = WSM::get_total_bytes_needed(nlevi_packs, 13+(n_wind_slots+n_trac_slots), policy);
+  const size_t wsm_request   = WSM::get_total_bytes_needed(nlevi_packs, 13+(n_wind_slots+n_trac_slots), policy);
 
   return interface_request + wsm_request;
 }
@@ -238,7 +238,7 @@ void SHOCMacrophysics::init_buffers(const ATMBufferManager &buffer_manager)
   const int wsm_size     = WSM::get_total_bytes_needed(nlevi_packs, 13+(n_wind_slots+n_trac_slots), policy)/sizeof(Spack);
   s_mem += wsm_size;
 
-  int used_mem = (reinterpret_cast<Real*>(s_mem) - buffer_manager.get_memory())*sizeof(Real);
+  size_t used_mem = (reinterpret_cast<Real*>(s_mem) - buffer_manager.get_memory())*sizeof(Real);
   EKAT_REQUIRE_MSG(used_mem==requested_buffer_size_in_bytes(), "Error! Used memory != requested memory for SHOCMacrophysics.");
 }
